# Decimal to floating point conversion