--- a/src/etc/test-float-parse/huge-pow10.rs
+++ b/src/etc/test-float-parse/huge-pow10.rs
@@ -1,0 +1,21 @@
+// Copyright 2015 The Rust Project Developers. See the COPYRIGHT
+// file at the top-level directory of this distribution and at
+// http://rust-lang.org/COPYRIGHT.
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+mod _common;
+
+use _common::validate;
+
+fn main() {
+    for e in 300..310 {
+        for i in 0..100000 {
+            validate(format!("{}e{}", i, e));
+        }
+    }
+}