--- a/src/libcoretest/num/mod.rs
+++ b/src/libcoretest/num/mod.rs
@@ -30,6 +30,7 @@ mod u32;
 mod u64;
 
 mod flt2dec;
+mod dec2flt;
 
 /// Helper function for testing numeric operations
 pub fn test_num<T>(ten: T, two: T) where