… src/etc/dec2flt_table.py
… src/libcore/num/dec2flt/mod.rs
… src/libcore/num/dec2flt/parse.rs
… src/libcore/num/dec2flt/rawfp.rs
… src/libcore/num/dec2flt/table.rs
… src/libcore/num/flt2dec/bignum.rs
… src/libcore/num/flt2dec/strategy/grisu.rs
… src/libcore/num/mod.rs
… src/libcoretest/num/dec2flt/mod.rs
… src/libcoretest/num/dec2flt/parse.rs
… src/libcoretest/num/dec2flt/rawfp.rs
… src/libcoretest/num/flt2dec/bignum.rs
… src/libcoretest/num/flt2dec/strategy/dragon.rs