This PR completely rewrites the conversion of decimal strings to `f64` and `f32`. It is intended to be absolutely positively completely 100% accurate. To the best of my knowledge, it achieves that goal completely. Any input that is not rejected is converted to the floating point number that is closest to the true value of the input. This includes overflow, subnormal numbers, and underflow to zero. In other words, the rounding error is less than or equal to 0.5 units in the last place. Half-way cases (exactly 0.5 ULP error) are handled with half-to-even rounding (also known as banker's rounding).

This code implements the algorithms from the paper [How to Read Floating Point Numbers Accurately][paper] by William D. Clinger, but that only accounts for about 300 lines of code. A good portion of the code implements (often quite non-trivial) functions not described in the paper because they are built-in in Scheme, which is used to describe the algorithms. Furthermore, extensions (hinted at in the paper but with little detail) to the algorithms handle inputs that result in subnormal numbers, plus some algorithmic optimizations.

## Correctness

With such a large amount of tricky code, many bugs are to be expected. Indeed I have spent far more time debugging wrongly-rounded results and fixing the underlying cause than I spent transliterating the algorithms.

Extensive tests (taking in the order of hours to run through to completion) are included in `src/etc/test-float-parse`: Though exhaustively testing all possible inputs is impossible, I've had good success with generating millions of instances from various "classes" of inputs. These have found most of the bugs I've fixed and will be useful to anyone modifying this code in the future. The sheer extent of the test suite is justified by the fact that I've had quite a few bugs that were only triggered by a small subset of the numbers in one particular class.

These tests take far too long to be run by @bors so contributors who touch this code need the discipline to run them. For @bors, there are small `#[test]`s, mostly composed of inputs that the aforementioned big test suite generated which uncovered bugs at some point. It doesn't cover every stupid mistake I made in course of writing this, and it won't cover every stupid mistake anyone (including me) could make in the future, but maybe it helps a little bit.

Another aspect is *integer* overflow. Extreme (or malicious) inputs can cause overflow both in the machine-sized integers used for bookkeeping throughout the algorithms (e.g., the decimal exponent) as well as the arbitrary-precision arithmetic that is necessary in edge cases (close to zero or infinity) and the occasional modestly-sized "tricky" input. I've spent a good while adding input validation to reject all such cases I know of, and I am quite sure nobody will *accidentally* cause this code to go out of range. Still, no guarantees.

## Limitations

Noticed the weasel words "Any input *that is not rejected*" at the beginning? Some otherwise well-formed decimal strings are rejected because:

- The decimal exponent is too big. This limitation is shared with the old code, though it accepts any exponent that fits into an i64 while this version stops at `i64::MAX / 10`, since it needs to do arithmetic with the exponent.
- So many digits that we can't even reliably do the necessary computations to check if it's zero or infinity. The bound for this is about 10^18 characters before and after the decimal point. Clearly this is a very serious regression that will impact almost all users.
- The value of the input requires too many digits, i.e., `digits * 10^abs(exp)` can't be stored in a bignum. This only applies if the value is not "obviously" zero or infinite, i.e., if you take a near-infinity or near-zero value and add many pointless fractional digits. To compute the precise value would require computing the true value as a fraction, which would overflow.

None of these limits are an issue for any realistic inputs. Still, it is technically possible to violate any of these conditions with a string that, when evaluated with unlimited precision, would result in an in-range float (e.g., an `n`-digit integer with an exponent of `-n`). I'm not losing any sleep over that, especially since the old code completely falls apart in those cases. Still, it was `Ok` before and is `Err` now, so technically this could break some programs.

However, the third bullet point would have rejected some sensible inputs, such as `1.7976931348623157e-324` (the smallest positive subnormal number, represented with 17 fractional digits; note that the same float is usually expressed as `5e-324`). With a conservative safety margin, the default of 36 * 32 bit was not sufficient to represent this number. I increased the size to 32 * 40 bit, which allows 370 decimal digits with the same safety margin as before, covering subnormals with 17 digits precision and much more.

This change affects the `flt2dec` code, specifically the Dragon4 algorithm. Except it doesn't -- as the comments in `bignum.rs` guessed, increasing the upper limit did not change performance at all. I ran the benchmarks in `coretest::num::flt2dec::strategy::dragon` and the effects of this change were mostly way below the noise threshold. `bench_big_exact_12` *sometimes* takes 11µs instead of 3µs I measured before applying this change, but at other times it does hit the 3µs. I don't want to revert the change just to see if this also happened before my change, so I'll just assume it did and chalk it up to the super noisy environment I ran it in.

## Performance

Benchmarks are included. The results are... surprisingly good, given that I spent little (though non-zero) time optimizing. Short numbers that hit the fast paths (`f64` multiplication or shortcuts to zero/inf) have comparable performance, sometimes a bit faster, sometimes a bit slower, but rarely more than a factor of two difference. We're talking tens of nanoseconds here.

Numbers that are delegated to Algorithm Bellerophon (using floats with 64 bit significand, implemented in software) are slower, but not drastically so. On my machine this takes a couple hundred nanoseconds.

Numbers that need the AlgorithmM fallback (for `f64`, roughly everything below 1e-305 and above 1e305) take far, far longer, hundreds of microseconds. Note that my implementation is not quite as naive as the expository version in the paper (it needs one to four division instead of 300 plus), but division is fundamentally pretty expensive and my implementation of it is extremely simple and slow.

All benchmarks run on a mediocre laptop with a i5-4200U CPU under light load.

## Binary size

Unfortunately the implementation needs to duplicate almost all code: Once for `f32` and once for `f64`. No, this cannot be avoided, at least not completely. See "Future work" below. There's also a precomputed table of powers of ten, weighing in at about six kilobytes.

Running a stage1 `rustc` over a stand-alone program that simply parses pi to `f32` and `f64` and outputs both results reveals that the overhead is about 44 KiB normally and about 28 KiB with LTO. It's presumably half of that + 3 KiB when only one of the two code paths is exercised.

| rustc options                 | old       | new       | delta         |
|---------------------------    |---------  |---------  |-----------    |
| [nothing]                     | 2588375   | 2633828   | 44.39 KiB     |
| -O                            | 2585211   | 2630688   | 44.41 KiB     |
| -O -C lto                     | 1026353   | 1054981   | 27.96 KiB     |
| -O -C lto -C link-args=-s     | 414208    | 442368    | 27.5 KiB      |

## Future Work

### Directory layout

This code heavily relies on two types introduced in PR #24612 (floating-to-decimal formatting), a bignum type and a software floating-point type with 64 bit significand and 16 bit exponent. This PR does not change the module structure, so currently `core::num::dec2flt` reaches into the innards of `core::num::flt2dec`, which is distasteful. I want to move the common code out of `flt2dec` into a upper-level module that makes it clear this code is shared by both conversion directions.

### Performance

I've spent a few hours applying micro optimizations to get to this point. Much more work could be done, especially for large inputs. Specific ideas are jotted down in `FIXME`s all over the code. Of particular note is the division algorithm, which I suspect (but for lack of profiling cannot prove) to be the primary bottleneck of AlgorithmM.

### Binary size

As described above, the code size is not as good as it could be. One could try to compress the table further, though I am skeptical. Another avenue would be reducing the code duplication from basically everything being generic over `T: RawFloat`. Perhaps one can reduce the magnitude of the duplication by pushing the parts that don't need to know the target type into separate functions, but this is finicky and probably makes some code read less naturally. It is left as future work, see below.

### Other bases

This PR leaves `f{32,64}::from_str_radix` alone. It only replaces `FromStr` (and thus `.parse()`). I am convinced that `from_str_radix` should not exist, and have proposed its [deprecation and speedy removal][deprecate-radix]. Whatever the outcome of that discussion, it is independent from, and out of scope for, this PR.

### Error enum

The error reporting for float parsing is currently *very* minimal. Empty strings get their own error message, but everything else is lumped in the grab bag "invalid". I don't know if this has a reason or if it's just a historic artifact, but the parsing code in this PR has enough context to give more detailed error messages. I'm not hell bent on changing `ParseFloatError[Kind]` but this is something one might want to do in the future.

Fixes #24557
Fixes #14353

r? @pnkfelix

cc @lifthrasiir @huonw 

[paper]: http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.45.4152
[deprecate-radix]: https://internals.rust-lang.org/t/deprecate-f-32-64-from-str-radix/2405