@eefriedman 2015-07-26
> The decimal exponent is too big. This limitation is shared with the old code, though it accepts any exponent that fits into an i64 while this version stops at i64::MAX / 10, since it needs to do arithmetic with the exponent.

This seems like something which could be handled trivially: if the exponent overflows, either the correct result is inf or 0, or the string has a stretch of zeros way too long to fit into memory.