@eefriedman 2015-07-26
> However, the third bullet point would have rejected some sensible inputs, such as 1.7976931348623157e-324 (the smallest positive subnormal number, represented with 17 fractional digits; note that the same float is usually expressed as 5e-324).

Umm, 1.7976931348623157e-324 ?  Are you sure you didn't mean some other number?