@eefriedman 2015-07-27
It would probably be a good idea to deprecate f32/f64::from_str_radix.  (We don't want to keep them around given that they don't actually work correctly.)