@eefriedman 2015-07-27
This might be a silly question... but does fast_path() work correctly on Linux x86-32 with SSE disabled?