@rkruppe 2015-07-27
> Umm, 1.7976931348623157e-324 ? Are you sure you didn't mean some other number?

Gaah. The fractional part belongs to another boundary case. Corrected.

> It would probably be a good idea to deprecate f32/f64::from_str_radix. (We don't want to keep them around given that they don't actually work correctly.)

Good to hear. I'd like to collect some more support before I go ahead and do it, and I'm not sure if it needs to go into this PR.

> This might be a silly question... but does fast_path() work correctly on Linux x86-32 with SSE disabled? (The Clinger paper calls out a similar scenario in section 9.)

Not silly at all! Unfortunately the VM on which I'd normally test this currently has... issues. If anyone could compile a short program doing a float multiplication with a 32 bit rust and `-C target-feature="-sse"` and post the assembly, that could be great. x87 *can* round correctly by setting the control word correctly, but GCC for example never does this. We'd have to see if LLVM allows doing so and if rustc instructs it to.