@Gankro 2015-07-27
re deprecation: kill it with fire