@pnkfelix 2015-08-06
Okay, done with review!  I didn't see anything major, and I trust @rkruppe will be good about addressing any of my nits as @rkruppe sees fit.

So, r=me once it is ready to land.  (I'll be out of town for a week, feel free to ping other people in `#rust-internals` for `r=pnkfelix` marker once its ready.)