@rkruppe 2015-08-06
So I finally got around to testing whether the fast path is correct on x86 without SSE (as pointed out by @eefriedman). Turns out the exact same problem described in the paper (in the context of a Motorola 68881/68882 floating point coprocessor) also applies to more common hardware. `1.448997445238699` is rounded incorrectly, to the float whose shortest decimal representation is `1.4489974452386991` (off by one bit). Bummer. I'll do *something* about this before merging, but I don't know yet *what*.