@rkruppe 2015-08-08
Adressed nits (except changing external tests to output ASCII, still in the works) and rebased.