@rkruppe 2015-08-11
Now that PR #27647 has been merged and the fast path brokenness is at least somewhat addressed (see last commit), I declare this PR officially **ready**.