@eefriedman 2015-08-11
Isn't `(target_arch="x86", target_os="ios")` targeting the iOS simulator?  In that case, the fact that it doesn't default to SSE2 enabled is probably an accident.