@rkruppe 2015-08-11
Yes, it's targeting the iOS simulator. Since @cmr brought that up, but later called the fact that x86 MSVC does not use SSE an accident, I can only assume it's *not* an accident.