@rkruppe 2015-08-12
Since there's no in-tree target that disables SSE any more, "deal with it" now means "mention that it's theoretically broken and add a test that will expose said brokenness if some unfortunate soul is forced to compile without SSE".