@IvanUkhov 2015-08-14
@rkruppe, thank you for this work!